Add alert text for wider terminals
