+++ alerta/top.py
@@ -15,10 +15,72 @@ from datetime import datetime
 
 from alerta.api import ApiClient
 from alerta.alert import AlertDocument
-from alerta.utils import Helpers
 
 SCREEN_REDRAW_INTERVAL = 2
 lock = threading.Lock()
+
+
+CRITICAL_SEV_CODE = 1
+MAJOR_SEV_CODE = 2
+MINOR_SEV_CODE = 3
+WARNING_SEV_CODE = 4
+INDETER_SEV_CODE = 5
+CLEARED_SEV_CODE = 5
+NORMAL_SEV_CODE = 5
+OK_SEV_CODE = 5
+INFORM_SEV_CODE = 6
+DEBUG_SEV_CODE = 7
+AUTH_SEV_CODE = 8
+UNKNOWN_SEV_CODE = 9
+
+# NOTE: The display text in single quotes can be changed depending on preference.
+# eg. CRITICAL = 'critical' or CRITICAL = 'CRITICAL'
+
+CRITICAL = 'critical'
+MAJOR = 'major'
+MINOR = 'minor'
+WARNING = 'warning'
+INDETERMINATE = 'indeterminate'
+CLEARED = 'cleared'
+NORMAL = 'normal'
+OK = 'ok'
+INFORM = 'informational'
+DEBUG = 'debug'
+AUTH = 'security'
+UNKNOWN = 'unknown'
+NOT_VALID = 'notValid'
+
+ALL = [CRITICAL, MAJOR, MINOR, WARNING, INDETERMINATE, CLEARED, NORMAL, OK, INFORM, DEBUG, AUTH, UNKNOWN, NOT_VALID]
+
+_SEVERITY_MAP = {
+    CRITICAL: CRITICAL_SEV_CODE,
+    MAJOR: MAJOR_SEV_CODE,
+    MINOR: MINOR_SEV_CODE,
+    WARNING: WARNING_SEV_CODE,
+    INDETERMINATE: INDETER_SEV_CODE,
+    CLEARED: CLEARED_SEV_CODE,
+    NORMAL: NORMAL_SEV_CODE,
+    OK: OK_SEV_CODE,
+    INFORM: INFORM_SEV_CODE,
+    DEBUG: DEBUG_SEV_CODE,
+    AUTH: AUTH_SEV_CODE,
+    UNKNOWN: UNKNOWN_SEV_CODE,
+}
+
+_DISPLAY_SEVERITY = {
+    CRITICAL:      "Crit",
+    MAJOR:         "Majr",
+    MINOR:         "Minr",
+    WARNING:       "Warn",
+    INDETERMINATE: "Ind ",
+    CLEARED:       "Clr",
+    NORMAL:        "Norm",
+    INFORM:        "Info",
+    OK:            "Ok",
+    DEBUG:         "Dbug",
+    AUTH:          "Sec",
+    UNKNOWN:       "Unkn"
+}
 
 
 class UpdateThread(threading.Thread):
@@ -70,7 +132,7 @@ class UpdateThread(threading.Thread):
 
         while self.running:
             self.status = 'updating...'
-            from_date = self._update()
+            from_date = self._update(from_date)
             try:
                 time.sleep(self.interval)
             except (KeyboardInterrupt, SystemExit):
@@ -120,22 +182,16 @@ class UpdateThread(threading.Thread):
 
         self.app_last_time = app_time
 
-        query = dict()
-        if from_date:
-            query['from-date'] = from_date
-
-        if 'sort-by' not in query:
-            query['sort-by'] = 'lastReceiveTime'
-
+        query = {'from-date': from_date}
         response = self.api.get_alerts(query)
 
         with lock:
             self.status = '%s - %s' % (response['status'], response.get('message', 'no errors'))
             self.total = response.get('total', 0)
-            self.alerts = response.get('alerts', [])
+            alert_delta = response.get('alerts', [])
             self.last_time = datetime.strptime(response.get('lastTime', ''), "%Y-%m-%dT%H:%M:%S.%fZ")
 
-            for alert in self.alerts:
+            for alert in alert_delta:
 
                 key = (alert['environment'], alert['resource'])
                 if key not in self.resources:
@@ -159,6 +215,11 @@ class UpdateThread(threading.Thread):
                     self.origins[key]['count'] = 0
                 self.origins[key]['count'] += 1
                 self.origins[key]['origin'] = alert['origin']
+
+        response = self.api.get_alerts({})
+        with lock:
+            self.alerts = response.get('alerts', [])
+            self.total = response.get('total', 0)
 
         return response.get('lastTime', '')
 
@@ -383,24 +444,52 @@ class Screen(object):
                     if i == self.max_y - 3:
                         break
             else:
-                self._addstr(2, 1, 'Sev. Time     Dupl. Env.        Service    Resource     Event        Value', curses.A_LOW)
+                if self.max_x < 100:
+                    self._addstr(2, 1, 'Sev. Time     Dupl. Env.         Service      Resource     Event        Value ',
+                                 curses.A_UNDERLINE)
+                else:
+                    self._addstr(2, 1, 'Sev. Last Recv. Time Dupl. Customer Environment  Service      Resource ' +
+                                       '      Event          Value   Text' + ' ' * (self.max_x - 106), curses.A_UNDERLINE)
+
+                def name_to_code(name):
+                    return _SEVERITY_MAP.get(name.lower(), UNKNOWN_SEV_CODE)
+
+                def to_short_sev(severity):
+                    return _DISPLAY_SEVERITY.get(severity, _DISPLAY_SEVERITY['unknown'])
+
                 i = 2
+                #self._addstr(1,1, '%sx%s' % (self.max_y, self.max_x))
                 alerts = copy.deepcopy(self.w.alerts)
-                for alert in sorted(alerts, key=lambda x: Helpers.name_to_code(x['severity'])):
+                for alert in sorted(alerts, key=lambda x: name_to_code(x['severity'])):
                     a = AlertDocument.parse_alert(alert)
                     i += 1
                     color = self._COLOR_MAP.get(a.severity, self._COLOR_MAP['unknown'])
                     try:
-                        self._addstr(i, 1, '{0} {1} {2:5d} {3:<11} {4:<10} {5:<12.12} {6:<12.12} {7:<5.5}'.format(
-                            Helpers.to_short_sev(a.severity),
-                            a.get_date('last_receive_time', 'short')[7:],
-                            a.duplicate_count,
-                            a.environment,
-                            ','.join(a.service),
-                            a.resource,
-                            a.event,
-                            a.value
-                        ), color)
+                        if self.max_x < 100:
+                            self._addstr(i, 1, '{0:<4} {1} {2:5d} {3:<12} {4:<12} {5:<12.12} {6:<12.12} {7:<6.6}'.format(
+                                to_short_sev(a.severity),
+                                a.get_date('last_receive_time', 'short')[7:],
+                                a.duplicate_count,
+                                a.environment,
+                                ','.join(a.service),
+                                a.resource,
+                                a.event,
+                                a.value
+                            ), color)
+                        else:
+                            self._addstr(i, 1, '{0:<4} {1} {2:5d} {3:>8.8} {4:<12} {5:<12} {6:<14.14} {7:<14.14} {8:<7.7} {9:{10}.{10}}'.format(
+                                to_short_sev(a.severity),
+                                a.get_date('last_receive_time', 'short'),
+                                a.duplicate_count,
+                                a.customer or '-',
+                                a.environment,
+                                ','.join(a.service),
+                                a.resource,
+                                a.event,
+                                a.value,
+                                a.text,
+                                self.max_x - 102
+                            ), color)
                     except curses.error as e:
                         print str(e)
                         sys.exit(1)
